SERVER-27209 Propagate removal of BSONElement::valuestr()
- Fix: Remove `valuestr()` to remove potentially dangerous handling of embedded NULLs in string fields

- Fix: Change signedness of size parameter to unsigned

- Fix: Change callsites of `BSONObj::getStringField()` to account for changed return type `char*` -> `StringData`

Related PR in server repo: [SERVER-27209 Eliminate dangerous BSONElement string extraction methods](https://github.com/10gen/mongo/pull/2579)

GitOrigin-RevId: 65f2a1b1ba3bfacac7d349aae6391563ab3b4702

+++ src/mongo/db/modules/enterprise/src/snmp/serverstatus_client.h
@@ -29,7 +29,7 @@ public:
     int64_t getInt64Field(StringData name);
     void getInt64FieldAsString(StringData name, char* o_value, int o_valueLen);
     void getDoubleField(StringData name, char* o_value, int o_valueLen);
-    void getStringField(StringData name, char* o_value, int o_valueLen);
+    void getStringField(StringData name, char* o_value, size_t o_valueLen);
     void getDateField(StringData name, char* o_value, int o_valueLen);
 
     static const std::string NO_EXTRA;
